% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundance_consumers}
\alias{abundance_consumers}
\title{Total abundance of herbivores and detritivores}
\format{
Numeric variable with 68 observations
}
\source{
{Experimental Data}
}
\usage{
abundance_consumers
}
\description{
total abundance of herbivores and detritivores  considering all surveys.
}
\keyword{datasets}
