% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundance_predators}
\alias{abundance_predators}
\title{Total abundance of predatory insects}
\format{
Numeric variable with 68 observations
}
\source{
{Experimental Data}
}
\usage{
abundance_predators
}
\description{
total abundance of predatory insects considering all surveys.
}
\keyword{datasets}
