% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com}
\alias{com}
\title{Taxa abundances for all surveys}
\format{
A data frame with 68 rows or 'observations' and 20 variables or 'taxa':
}
\source{
{Experimental Data}
}
\usage{
com
}
\description{
A dataset containing the abundances of the taxa that were present in more than three observations in all surveys.
}
\keyword{datasets}
