% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_orig}
\alias{com_orig}
\title{All taxa abundances for all surveys}
\format{
A data frame with 68 rows or 'observarions' and 37 variables, the last 36 being the sampled 'taxa':
\describe{
\item{ID}{The identity code of the pond}
...
}
}
\source{
{Experimental Data}
}
\usage{
com_orig
}
\description{
A dataset containing the abundances of all taxa in all surveys.
}
\keyword{datasets}
