% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS1}
\alias{com_SS1}
\title{Taxa abundances for the first survey}
\format{
A data frame with 24 rows or 'observations' and 9 variables or taxa':
}
\source{
{Experimental Data}
}
\usage{
com_SS1
}
\description{
A dataset containing the abundances of the taxa that were present in more than two observations in the FIRST survey.
}
\keyword{datasets}
