% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{com_SS2}
\alias{com_SS2}
\title{Taxa abundances for the second survey}
\format{
A data frame with 24 rows or 'observations' and 18 variables or taxa':
}
\source{
{Experimental Data}
}
\usage{
com_SS2
}
\description{
A dataset containing the abundances of the taxa that were present in more than two observations in the SECOND survey.
}
\keyword{datasets}
