% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dif_dist_function.R
\name{Dif_dist}
\alias{Dif_dist}
\title{Difference in distance between centroids}
\usage{
Dif_dist(
  com,
  x1,
  x2,
  LVs = NULL,
  test = TRUE,
  nperm,
  family,
  num.lv = 2,
  strata = T,
  show.perm = T,
  orig_n.init = 30,
  perm_n.init = 1,
  type = "centroid",
  method = "VA",
  refit_perm = F
)
}
\arguments{
\item{x1}{A vector containing the identity of fish treatments.}

\item{x2}{A vector containing the identity of isolation treatments.}

\item{LVs}{The latent variables or ordination axis to be used to compute centroids. If NULL, GLLVM will be computed. Defaut is NULL.}

\item{test}{Should the significance of the difference between distances be tested? Defaut is set to TRUE. If FALSE, all arguments related to test will be ignored.}

\item{nperm}{The number of permutations.}

\item{family}{Family distribution to fit the GLLVMs.}

\item{num.lv}{Number of latent variables used to compute GLLVMs.}

\item{strata}{Should communities be permuted freely or be restricted by fish treatment? Defaut to TRUE.}

\item{show.perm}{Show permutation steps?}

\item{orig_n.init}{Number of iterations used to compute the original GLLVM? Defaut to 30.}

\item{perm_n.init}{Number of iterations used to compute the permuted GLLVMs? Defaut to 1.}

\item{type}{Either "centroid" or "median". Should centroids or spatial medians be computed?}

\item{method}{Either "VA" or "LA". The method used to estimate latent variables. See function gllvm from package gllvm}

\item{refit_perm}{If TRUE, the rows of the original community matrix will be permuted and for every permutation a gllvm will be computed. It can take too much time. If false, only the rows of a matrix with the original latent variables are permuted.}

\item{comm}{The community data. A species by site matrix containing abundance data.}
}
\description{
This function computes distances between centroids of ponds with and without fish for each isolation distance treatment. Then it computes the differences among those distances among isolation treatments and tests it for significance. IMPORTANT: This function will only work for the variables provided in this package, you may need to make changes to it if you want to apply it to a different data set.
}
