% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish Treatment identities for all Surveys}
\format{
Factor w/ 2 levels: "absent","present", and 68 observations
}
\source{
{Experimental Data}
}
\usage{
fish
}
\description{
Fish Treatment identities for all Surveys
}
\keyword{datasets}
