% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_isolation_SS1}
\alias{fish_isolation_SS1}
\title{Fish and Isolation Treatment identities in the First Survey}
\format{
Factor w/ 6 levels: ""30 absent", "120 absent", "480 absent", "30 present", "120 present", "480 present", and 24 observations
}
\source{
{Experimental Data}
}
\usage{
fish_isolation_SS1
}
\description{
Fish and Isolation Treatment identities in the First Survey
}
\keyword{datasets}
