% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish_SS2}
\alias{fish_SS2}
\title{Fish Treatment identities in the Second Survey}
\format{
Factor w/ 2 levels: "absent","present", and 24 observations
}
\source{
{Experimental Data}
}
\usage{
fish_SS2
}
\description{
Fish Treatment identities in the Second Survey
}
\keyword{datasets}
