% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isolation}
\alias{isolation}
\title{Isolation Treatment identities for all Surveys}
\format{
Factor w/ 3 levels: "30","120","480", and 68 observations
}
\source{
{Experimental Data}
}
\usage{
isolation
}
\description{
Isolation Treatment identities for all Surveys
}
\keyword{datasets}
