% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isolation_SS2}
\alias{isolation_SS2}
\title{Isolation Treatment identities in the Second Survey}
\format{
Factor w/ 3 levels: "30","120","480", and 24 observations
}
\source{
{Experimental Data}
}
\usage{
isolation_SS2
}
\description{
Isolation Treatment identities in the Second Survey
}
\keyword{datasets}
