% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isolation_SS3}
\alias{isolation_SS3}
\title{Isolation Treatment identities in the Third Survey}
\format{
Factor w/ 3 levels: "30","120","480", and 20 observations
}
\source{
{Experimental Data}
}
\usage{
isolation_SS3
}
\description{
Isolation Treatment identities in the Third Survey
}
\keyword{datasets}
