% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_coefplot_function.R
\name{My_coefplot}
\alias{My_coefplot}
\title{Plotting coefficients}
\usage{
My_coefplot(
  mles,
  upper,
  lower,
  species_labels = NULL,
  xlab = NULL,
  cex.axis = 1,
  y_spa = 0,
  rect = F,
  rect_lim = 5,
  ...
)
}
\arguments{
\item{mles}{A vector containing the maximum likelihood estimates to be plotted.}

\item{upper}{A vector containing the upper limits of the confidence intervals.}

\item{lower}{A vector containing the lower limits of the confidence intervals.}

\item{species_labels}{A vector containing the labels of each species.}

\item{xlab}{A label for the x axis.}

\item{cex.axis}{The size of the x axis.}

\item{y_spa}{Space to be added to the minimum and maximum values. This is to improve visualization.}

\item{rect}{Should assign different backgrounds to predators and non-predators? Default is to FALSE.}

\item{rect_lim}{Limit of the background.}

\item{...}{Other graphical parameters.}
}
\description{
This function is to plot the maximum likelihood estimates and their respective confidence intervals.
}
