% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelinson.et.al.2020-package.R
\docType{package}
\name{pelinson.et.al.2020-package}
\alias{pelinson.et.al.2020}
\alias{pelinson.et.al.2020-package}
\title{pelinson.et.al.2020: Data Analysis of Pelinson et al 2020}
\description{
This package is intended to reproduce the data analysis presented
    by Pelinson et al 2020.
}
\author{
\strong{Maintainer}: Rodolfo Pelinson \email{rodolfopelinson@gmail.com} (\href{https://orcid.org/0000-0002-8894-588X}{ORCID})

}
\keyword{internal}
