% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelinson.et.al.2020-package.R
\docType{package}
\name{Pelinson.et.al.2020A-package}
\alias{Pelinson.et.al.2020A}
\alias{Pelinson.et.al.2020A-package}
\title{Pelinson.et.al.2020A: Data Analysis of Pelinson et al 2020}
\description{
This package is intended to reproduce the data analysis presented
    by Pelinson et al 2020.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RodolfoPelinson/Pelinson.et.al.2020A}
  \item Report bugs at \url{https://github.com/RodolfoPelinson/Pelinson.et.al.2020A/issues}
}

}
\author{
\strong{Maintainer}: Rodolfo Pelinson \email{rodolfopelinson@gmail.com} (\href{https://orcid.org/0000-0002-8894-588X}{ORCID})

}
\keyword{internal}
