% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_null_function.R
\name{plot_null}
\alias{plot_null}
\title{Plotting null ordinations}
\usage{
plot_null(
  com,
  x1,
  x2,
  nperm = 1,
  family,
  num.lv = 2,
  strata = F,
  orig_n.init = 1,
  perm_n.init = 1,
  refit_perm = F,
  ...
)
}
\arguments{
\item{x1}{A vector containing the identity of fish treatments.}

\item{x2}{A vector containing the identity of isolation treatments.}

\item{nperm}{The number of permutations.}

\item{family}{Family distribution to fit the GLLVMs.}

\item{num.lv}{Number of latent variables used to compute GLLVMs.}

\item{strata}{Should communities be permuted freely or be restricted by fish treatment? Default to TRUE.}

\item{orig_n.init}{Number of iterations used to compute the original GLLVM? Default to 30.}

\item{perm_n.init}{Number of iterations used to compute the permuted GLLVMs? Default to 1.}

\item{refit_perm}{If TRUE, the rows of the original community matrix will be permuted and for every permutation a gllvm will be computed. It can take too much time. If false, only the rows of a matrix with the original latent variables are permuted.}

\item{comm}{The community data. A species by site matrix containing abundance data.}
}
\description{
This function is just for visualization of null generated communities. It is useful to see if your permutation design is doing what it is suppose to! It relies on the plot_ordination function.
}
