% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ordination_function.R
\name{plot_ordination}
\alias{plot_ordination}
\title{Plotting ordinations}
\usage{
plot_ordination(
  model_gllvm = NULL,
  LVs = NULL,
  species_loadings = NULL,
  xlim = NULL,
  ylim = NULL,
  x1,
  x2 = NULL,
  main = NULL,
  species = F,
  species_col = "black",
  name_species_col = "white",
  species_size = 4,
  sites_size = 1,
  species_names = NULL,
  sites = T,
  elipse = F,
  site_colors = NULL,
  pch = NULL,
  legend_labels = NULL,
  name_species_size = 0.8,
  legend_x = NULL,
  legend_y = NULL,
  legend = F,
  legend_horiz = F,
  legend_ncol = 1,
  ylab = "Latent Variable 2",
  xlab = "Latent Variable 1",
  alpha.elipse = 75,
  sep = F
)
}
\arguments{
\item{model_gllvm}{The resultant object from function gllvm.}

\item{LVs}{The latent variables to be plotted. If provided, the model_gllvm argument will be disregarded.}

\item{species_loadings}{A vector containing the species ladings.}

\item{xlim}{A vector of length = 2 with upper and lower limits of the x axis}

\item{x1}{First grouping variable}

\item{x2}{Second grouping variable}

\item{main}{Main Title}

\item{species}{Should species be plotted?}

\item{species_col}{Species colors. Either a color or a vector of colors with length equal to the number of species.}

\item{species_size}{Species sizes Either a color or a vector of colors with length equal to the number of sites.}

\item{sites_size}{Sites sizes. Either a number or a vector of numbers with length equal to the number of sites.}

\item{species_names}{Species names to be plotted. A vector of the names of the species.}

\item{sites}{Should sites be plotted?}

\item{elipse}{Should ellipses around treatments be plotted?}

\item{site_colors}{Sites colors. Either a color or a vector of colors with length equal to the number of sites.}

\item{pch}{Tipe of dots to plotted for sites.}

\item{legend_labels}{Labels for the legend.}

\item{name_species_size}{A cex value for species names. Either a number or a vector of numbers.}

\item{legend_x}{x coordinates to plot the legend}

\item{legend_y}{x coordinates to plot the legend}

\item{legend}{Should legend be plotted? Default to F.}

\item{legend_horiz}{Should legend be horizontal?}

\item{legend_ncol}{Number of columns in the legend.}

\item{ylab}{Label of y axis.}

\item{xlab}{Label of x axis.}

\item{alpha.elipse}{Transparency for the ellipses.}

\item{sep}{Should plots be separated by isolation distance to improve visualization?}

\item{ylimy}{A vector of length = 2 with upper and lower limits of the x axis}
}
\description{
This function is to plot the ordinations from the GLLVMs.
}
