% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_value_function.R
\name{sr_value}
\alias{sr_value}
\title{Plotting coefficients}
\usage{
sr_value(
  dfxy,
  z,
  xax = 1,
  yax = 2,
  method = c("bubble", "greylevel"),
  zmax = NULL,
  csize = 1,
  cpoint = 0,
  pch = 20,
  clegend = 0.75,
  neig = NULL,
  cneig = 1,
  xlim = NULL,
  ylim = NULL,
  grid = TRUE,
  addaxes = TRUE,
  cgrid = 0.75,
  include.origin = TRUE,
  origin = c(0, 0),
  sub = "",
  csub = 1,
  possub = "topleft",
  pixmap = NULL,
  contour = NULL,
  area = NULL,
  add.plot = FALSE
)
}
\description{
Slightly modified version of ade4's s.value graphical function. Draws round instead of square bubbles in some plots when argument "bubble" is called.
Author of the original function s.value: Daniel Chessel. Modification: Francois Gillet, 25 August 2012.
}
