% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey}
\alias{survey}
\title{Sampling Survey identities}
\format{
Factor w/ 3 levels: "1","2","3", and 68 observations
}
\source{
{Experimental Data}
}
\usage{
survey
}
\description{
Sampling Survey identities
}
\keyword{datasets}
