% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatments}
\alias{treatments}
\title{Explanatory variables}
\format{
A data frame with 68 rows or 'observations' and 4 variables:
\describe{
\item{ID}{The identity code of the pond}
\item{survey}{Number of the survey, '1' is the first, '2' is the second and '3' is the third}
\item{fish}{Fish treatment. Fish were either 'present' or 'absent'}
\item{isolation}{The isolation treatment. It could be either '30', '120' or '480'}
}
}
\source{
{Experimental Data}
}
\usage{
treatments
}
\description{
A dataset containing the explanatary variables for the experiment.
}
\keyword{datasets}
